+++ src/ReadoutEquipmentRORC.cxx
@@ -68,6 +68,7 @@ private:
       0; // flag to enable filling page with zeros before giving for writing
 
   int cfgFirmwareCheckEnabled = 1; // RORC lib check self-compatibility with fw
+  int cfgDebugStatsEnabled = 0;    // collect and print more buffer stats
 
   unsigned long long statsRdhCheckOk =
       0; // number of RDH structs which have passed check ok
@@ -199,6 +200,12 @@ ReadoutEquipmentRORC::ReadoutEquipmentRORC(ConfigFile &cfg, std::string name)
       theLog.log(InfoLogger::Severity::Warning,
                  "Bypassing RORC firmware compatibility check");
     }
+
+    // configuration parameter: | equipment-rorc-* | debugStatsEnabled | int | 0
+    // | If set, enable extra statistics about internal buffers status. (printed
+    // to stdout when stopping) |
+    cfg.getOptionalValue<int>(name + ".debugStatsEnabled",
+                              cfgDebugStatsEnabled);
 
     // configuration parameter: | equipment-rorc-* | TFperiod | int | 256 |
     // Duration of a timeframe, in number of LHC orbits. |
@@ -693,6 +700,13 @@ void ReadoutEquipmentRORC::setDataOn() {
     if (RocFifoSize == 0) {
       RocFifoSize = 1;
     }
+    // enable enhanced statistics
+    if (cfgDebugStatsEnabled) {
+      equipmentStats[EquipmentStatsIndexes::fifoOccupancyFreeBlocks]
+          .enableHistogram(12, 0, RocFifoSize, 0);
+      equipmentStats[EquipmentStatsIndexes::fifoOccupancyReadyBlocks]
+          .enableHistogram(12, 0, RocFifoSize, 0);
+    }
   }
   ReadoutEquipment::setDataOn();
 }
@@ -752,5 +766,54 @@ void ReadoutEquipmentRORC::finalCounters() {
     theLog.log("Equipment %s : %llu pages (+ %llu lost + %llu empty)",
                name.c_str(), statsNumberOfPages, statsNumberOfPagesLost,
                statsNumberOfPagesEmpty);
+  }
+
+  if (cfgDebugStatsEnabled) {
+
+    printf("\n*** begin debug stats ***\n");
+
+    std::vector<double> tx;
+    std::vector<CounterValue> tv;
+    CounterValue ts;
+
+    auto dumpStats = [&](bool revert) {
+      ts = 0;
+      for (unsigned int i = 0; i < tx.size(); i++) {
+        ts += tv[i];
+      }
+      printf("Fifo used (%%)\tSamples count\tSamples fraction (%%)\n");
+      for (unsigned int i = 0; i < tx.size(); i++) {
+        double t1 = tx[i] * 100.0 / RocFifoSize;
+        if (revert) {
+          t1 = 100 - t1;
+        }
+        double tr = 0.0;
+        if (ts != 0) {
+          tr = tv[i] * 100.0 / ts;
+        }
+        if ((i == 0) || (i == tx.size() - 1)) {
+          printf("%3d       \t%13lu\t%3.1lf\n", (int)t1, tv[i], tr);
+        } else {
+          double t2 = tx[i + 1] * 100.0 / RocFifoSize;
+          if (revert) {
+            t2 = 100 - t2;
+          }
+          printf("%3d - %3d     \t%13lu\t%3.1lf\n", (int)t1, (int)t2, tv[i],
+                 tr);
+        }
+      }
+    };
+
+    equipmentStats[EquipmentStatsIndexes::fifoOccupancyFreeBlocks].getHisto(tx,
+                                                                            tv);
+    printf("\nRORC transfer queue\n");
+    dumpStats(1);
+
+    equipmentStats[EquipmentStatsIndexes::fifoOccupancyReadyBlocks].getHisto(
+        tx, tv);
+    printf("\nRORC ready queue\n");
+    dumpStats(0);
+
+    printf("\n*** end debug stats ***\n");
   }
 }